Fix docker multi-stage build
Previously `npm run start` was also building the application. Thereby
overwriting the output of the build-stage. This was fixed by spliting
the build and run scripts.
Additionally many files not needed to run the application where included
in the final image. This was minimized by only copying the required
files for every stage.